// src/types/api.ts
export interface ImageResponse {
  id: string;
  imageUrl: string;
  publicId?: string;
}

export interface ProductRequest {
  productName: string;
  slug: string;
  genericName: string;
  brand: string;
  composition: string;
  strength: string;
  dosageForm: string;
  description: string;
  categoryId: string;
  imageUrls?: string[];
}

export interface ProductResponse {
  id: string;
  productName: string;
  slug: string;
  genericName: string;
  brand: string;
  composition: string;
  strength: string;
  dosageForm: string;
  description: string;
  categoryId: string;
  categoryName: string;
  productImages: ImageResponse[];
}